dom/media/test/test_streams_element_capture_createObjectURL.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test that a MediaStream captured from one element plays back in another</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

var manager = new MediaTestManager;

function checkDrawImage(vout) {
  var canvas = document.createElement("canvas");
  var ctx = canvas.getContext("2d");
  ctx.drawImage(vout, 0, 0);
  var imgData = ctx.getImageData(0, 0, 1, 1);
  is(imgData.data[3], 255, "Check video frame pixel has been drawn");
}

function isGreaterThanOrEqualEps(a, b, msg) {
  ok(a >= b - 0.01,
     "Got " + a + ", expected at least " + b + "; " + msg);
}

function startTest(test, token) {
  manager.started(token);

  var v = document.createElement('video');
  var vout = document.createElement('video');
  vout.token = token;

  v.src = test.name;
  v.preload = "metadata"
  var stream;

  var checkEnded = function() {
    is(stream.currentTime, vout.currentTime, test.name + " stream final currentTime");
    if (test.duration) {
      isGreaterThanOrEqualEps(vout.currentTime, test.duration,
         test.name + " current time at end");
    }
    is(vout.readyState, vout.HAVE_CURRENT_DATA, test.name + " checking readyState");
    ok(vout.ended, test.name + " checking playback has ended");
    if (test.type.match(/^video/)) {
      checkDrawImage(vout);
    }
    vout.remove();
    URL.revokeObjectURL(vout.src);
    manager.finished(vout.token);
  };
  vout.addEventListener("ended", checkEnded);

  document.body.appendChild(vout);
  v.onloadedmetadata = function () {
    stream = v.mozCaptureStreamUntilEnded();
    is(stream.currentTime, 0, test.name + " stream initial currentTime");
    vout.src = URL.createObjectURL(stream);
    v.play();
    vout.play();
  };
}

SpecialPowers.pushPrefEnv(
  { "set": [["privacy.reduceTimerPrecision", false]]},
  function() {
    manager.runTests([getPlayableVideo(gSmallTests)], startTest);
  });
</script>
</pre>
</body>
</html>
